example-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile example-requirements.in
#
python-dotenv==1.0.1
    # via -r example-requirements.in
